requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
python-multipart==0.0.20
pdfplumber==0.11.8
pymongo==4.15.4
pydantic-settings==2.11.0
pytesseract==0.3.13
Pillow==12.1.1
spacy==3.8.11
nltk==3.9.2
en-core-web-sm==3.8.0
streamlit==1.51.0
requests==2.32.5
pandas==2.3.3
pytest==8.4.2
